Rename the lightningcss plugin to style
No sense tying this module name to LightningCSS. It could probably be
implemented using PostCSS or Sass.
