+++ eleventy.config.js
@@ -8,13 +8,13 @@ const anchor = require('markdown-it-anchor');
 const blob = require("./utils/blob.js");
 const collections = require('./utils/collections.js');
 const filters = require('./utils/filters.js');
-const lightningcss = require('./utils/lightningcss.js');
+const style = require('./utils/style.js');
 const shortcodes = require('./utils/shortcodes.js');
 const transforms = require('./utils/transforms.js');
 
 module.exports = function (config) {
   // Plugins
-  config.addPlugin(lightningcss);
+  config.addPlugin(style);
   config.addPlugin(pluginRss);
   config.addPlugin(syntaxHighlight, {
     templateFormats: ["njk", "md"],
